WIP - Changed how TaskManager works
There is now a readyList and everything in it gets executed in round-robin order. Delay spawns a thread which waits for the specified time and then puts the Task back in the readyList. This is very unsafe. First of all, The Task pointer isn't protected at all. Second, the delay thread is putting the Task in the readyList without any locks, so maybe it works and maybe it fails miserably. I need to implement a Mutex like I did Thread.

I'm also putting items in the waitEventList but never managing it