Added stack size to Thread constructor
This allows the main thread to have a large stack and delay threads to have a smaller one. Setting all stacks to 4K was causing delay thread to fail to spawn (presumably from running out of stack space). But lowering the stack size caused the main thread to run out of stack space