Changed SystemInterface (and other things like TaskManager) to not have global constructor
TaskManager was constructing too early so it crashed when it tried to spawn a thread. So I decided to have Application create a SystemInterface instance when it is constructed. This causes TaskManager to construct after the system is up and running. It's also a better design. I want to look into puttng Mallocator and Global (the other 2 singletons) into SystemInterface to get rid of their global ctors