Fixed upValue in myDelay
I had the same issue in mrsh. upValues are really passed by value. If the value is still open you will change it as though it were passed by reference. But when it gets closed it acts like it's passed by value. So to change an upValue you have to pass it as an array with one value and then change that. I need to figure out a way around this. I wonder how Lua does it.