Fixed expression handling
expression() is actually a list of arithmeticExpressions separated by commas. So I renamed it to commaExpression and then went through the code and changed all instances of expression() to use either arithmeticExpression() or commaExpression(). I followed the C++ BNF here, but basically I use commaExpression() whereever the comma separator would not be ambiguous.