Implemented native delay
The previous delay would just return Delay as the CallReturnValue which would cause the TaskManager to avoid executing the Task again until the delay time was over. But this didn't handle events coming in during the delay. So I got rid of all this TaskManager dealing with delays logic and tried an implementation written in m8rscript, where I spawn a thread which delays for the specified period and then sets a done flag. In the meantime I waitForEvent() and check the state of the done flag. When an event came in the waitForEvent would return but done would not be set so it would wait some more.

This all worked fine. So I did an implementation in native code. This pushes a dummy function on the EU stack so we can reenter safely and continue to wait. When the done flag is set I return from the dummy function and we continue normally. If an event comes in, it pushes its function onto the stack, executes it, then pops its function and we continue to wait for the delay time.