Fixed bug in handling LAND and LOR and made it more efficient
The operations were not handling chaining right. You can't do a dup() on a register and on the second test it was trying to do that. At the end of either side of the LAND test you simply need to put a 0 or 1 in the result register. So we really don't need LAND opcodes at all. I'll get rid of them at some point.