Added overridable value() method to Object
Currently just works with toStringValue.If Value is an Object, it first calls value() to see if it returns a defined value. If so, it does toString on that value. Otherwise it does what it used to do. This allows a getValue() m8rscript function to override the Value. It's currently used for iterator. So now if you try to convert the iterator object to a string it will call Iterator::getValue which will return the current iterated value.