Made it possible to add new properties to an Object
Saying foo.bar = 5 is an error is Object foo doesn't already contain a bar property. This is to avoid the common error of saying foo.Bar instead of foo.bar and getting 2 properties. To add a property you need to do foo["bar"] = 5, which is a more explicit syntax. But that got broken somewhere.