Fixed code generation error
I had an optimization that would try to save a byte when the pop result was discarded. It would look at the code and change a pop to a popx if the result wasn't used.The problem is it sometimes made a mistake. So for now I've removed the optimization until I figure out a better way to do it.