Changed setElement to take the same Type enum as setProperty
For setting elements you really need all 3 cases. NeverAdd says that unless the index exists, don't add it. AlwaysAdd says to ignore the passed index and always append if this is an Array. AddIfNeeded says to replace the indexed value if it exists. Otherwise extend the size of the array and add it. If setElement is used for MaterObject then AlwaysAdd and AddIfNeeded do the same thing - if the property exists replace its value, otherwise add it