Optimized ExecutionUnit dispatcher
It was doing about 4 tests and keeping a yield counter. Got rid of all that, replacing it with a single boolean _checkForExceptions. This gets set if anything like termination happens or too many errors, or events coming in, or a yield is requested. So now dispatch is just check tis flag and get the next opcode. Speeds things up by about 25%.

I got rid of the yield counter, which would yield after executing a certain number of instructions. This was not only more expensive, but it was not a good judge of runtime of the task. So I replaced it with a timer based mechanism in TaskManager. The time starts when the task starts executing. When it fires, the _yield flag in the EU is set which makes it yield on the next instruction.