Changed Mallocator to use malloc/free
The problem with using my own allocator is that I have to allocate a block of memory to use for it from the RTOS. That means I have to leave enough behind for other uses. This way I can use all of memory. Wifi uses a lot of memory and I'm down to 51K, but that's better than before where I was down to 41K.