Changed the run loop to make it compatible with Arduino
Arduino has no threading model, and adding one would be a hassle. So I have rearranged the general runloop to work the same as Arduino, where is is called repeatedly. I also got rid of all threading and changed Timer to work with an interrupt driver hardware timer.

SystemInterface now has startTimer/stopTimer functions.On ESP these use the hardware timer1 and when it fires the callback is called. This means it's being called in an ISR, so we're limited in what we can do. So it simply calls TaskManager::yield() which will cause any currently running ExecutionUnit to yield at which point the TaskManager checks any active timer and fires its callback in the main thread of code.

Application::runLoop is now Application::runOneIteration and it's called over and over again on Mac. On ESP it will get called repeatedly in the loop() function. This means I can get rid of Thread entirely, which should save space.

One problem still to solve is how to handle contention.Since everything is running in one thread now with an ISR that fires, we just have to protect the variables accessed by the ISR. So I will go back to a single global lock which on Mac will be a normal mutex and on ESP it will be a disable/enable interrupts