Changed how Error printing works
Error used to have print function which would be passed an ExecutionUnit and it would use the EU print function to do the printing. This is needed so printing to go to the correct console. But it put a dependency on the EU in Error. So I made Error have format functions that return a String and the caller prints to the print function of choice. This will make TCPServer cleaner