Fixed a bug with bogus captures.
Note: this is the same comment as 9c8b4f077a75645b531271c51f73fbe8db13ee6c. That comment was wrong. It was actually just an optimization in passing a lambda.

Capturing by & means that when the caller goes away the captured values become invalid. Made captures explicit so now they are value throughout the lifetime of the lambda. This solves a problem where the connectionID was bogus