Moved execution loop to a separate thread.
Now the main thread runs the WebView event loop, which is necessary for event handling. And  m8rScript execution is done in a background thread. If runOneIteration return false, meaning it didn't do anything, we wait 2ms for events to come in or timers to fire. This gets the CPU overhead down from 100% to 1%