Added ability to change default heartrate on time
An on time of 1ms works well for ESP because it keeps the light from blinking too brightly. But that doesn't show up on m8rsim because the refresh rate is too low. This will allow it to be bumped up to be visible for m8rsim while keeping the lower value for ESP