Switched to creating a mainTask to run the application loop
The startup task really shouldn't be used for any ongoing tasks. This allows me to set the stack size of the main task, which is where all the action is

Now the mem command runs without crashing. But it doesn't output its results until I hit return several times. Must be an issue with buffer sizes or something