Made heartbeat use startTimer/stopTimer directly
Using a Timer object for the heartbeat is problematic because having a 1ms off time is too short for the Timer to handle. This causes  the LED to be on for too long which makes it too bright.

Works on Rtos and Arduino, but stops working after a while on Mac. I  think it doesn't like restarting the timer in the timer callback.