WIP - Adding WiFi config setup
Currently we try to start the network from saved credentials. If not found it does a network scan, prints all the networks found and then starts an AP with the name (for now) "MyCoolESPAP". Ultimately I will start an HTTPServer on this AP and present the networks to be chosen and a password entered. Then WiFi will start back up with this network. For now to get a network primed you can set the SetupTestSTAConnection flag to true and it will connect to my network. Then you can turn this flag off and it should connect to that network from then on