Add table of special chars
This is generated by generateValues (used to be generateSharedAtoms). List is a string with one or two character sequences preceded by the corresponding token. Use strstr to find sequences in this list and do additional processing to handle 3 and 4 character sequences: <<=, >>=, >>>, >>>=. Saves about 300 bytes and makes Scanner cleaner and easier to modify which special character sequences are handled