Moved Request to be a public structure and completed parsing of header
Now the entire header is parsed, including params and the header key/value pairs. The RequestFunction now takes a Request so we can now write a function which will process, for instance, the URI params. Next step is to write a set of functions that implement a REST API for the Network setup page