Got rid of file handling from RequestHandler
The Request used to have both a function and a path. Both would never be set. If the path is set then when a matching URI is received the path would be opened and the file would be returned in the response. Now all RequestHandlers have a RequestFunction. When the on() call is made with a file path, a RequestFunction is generated which handles the file response.