Got rid of builtinVerbs map
The builtinVerbs correspond to SharedAtoms, which have numeric values less than ExternalAtomOffset. Value::Type places all the non-built-in verb enums starting at ExternalAtomOffset. So a Value::Type less than ExternalAtomOffset is a built-in verb. This avoids the map lookup and saves a lot of space.

Currently any identifier that matches a SharedAtom is added as a built-in verb which causes an error exit (currently without any error message) if a function for that  verb is not present.