Added ability to have enums that are keywords.
Having an enum like 'if' isn't possible because it's a keyword. To support this in the shared atoms, I added a '$' to the end of keywords in SharedAtoms.txt. When the generator see 'if$' it will make that the enum value, but will make the string 'if'. So you say SA::if$ and it will be the string 'if'