Added baseTime to SystemTime
The current time value returned is measured from the Epoch. In microseconds that a huge number. It fits in a uint64_t but not in a 64 bit Float. At the start of the program I save the currentTime and then subtract it. This probably breaks real time calcs, but that can be fixed later