Get rid of ROMString
ESP RTOS automatically puts all read-only data (strings and data arrays) into Flash. When you access that data on a non-aligned boundary, ESP RTOS has logic to do the equivalent of ROMReadByte(). So none of the ROMString stuff is needed. The consequence of all this is that Arduino is dead. That's fine since ESP RTOS is working really well and I'm planning on more changes that will depend on FreeRTOS