WIP - Got Marly working on Mac
Marly (and all scripting languages are now registered by main and an attempt is made to load when a scripting language is present. Works with Marly on Mac. Not yet including it on ESP. Also generating marly specific shared atoms, but not registering any of them so they are not being found during a Marly compile