Got Marly execution working and implemented delay
Now codeStack contains Values. At the start of execution it contains one value, the main program list. From then on it contains sets of 3 values: The current list, the index into that list and the current state. This will allow nesting of lists and the state will let us know how to manage the stack when we're done with the current list. Right now all the basic functionality works, but any loops don't