Put all tokens from libm8r into ParseEngine
This allows a unified Token for m8rscript parsing.

This required huge changes to lots of files because I got rid of 'using namespace m8r' in m8rscript headers. It's a bad idea to do this anyway and it makes it easier to distinguish m8rscript::Token from m8r::Token. This caused lots of changes, both in specifying the m8r namespace in m8rscript header files and moving definitions around to minimize having to use namespaces. I still have 'using namespace m8r' in many m8rscript source files which is fine.

Still fails to parse hello.m8r and I'm still not handling multi-char operators. But it builds and runs on Mac